new contract organization
