+++ programs/demo-tokens/src/instructions/mod.rs
@@ -1,0 +1,8 @@
+pub use initialize::*;
+pub mod initialize;
+
+pub use submit_token::*;
+pub mod submit_token;
+
+pub use token_vote::*;
+pub mod token_vote;
