+++ programs/demo-tokens/src/state/protocol.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+#[account]
+pub struct Protocol {
+    pub bump: u8,
+    pub num_tickers: u8,
+}
